darwin-aarch64: add support for MAP_JIT/pthread_jit_write_protect_np
