+++ substratevm/src/com.oracle.svm.core.genscavenge/src/com/oracle/svm/core/genscavenge/GCImpl.java
@@ -788,33 +788,40 @@ public final class GCImpl implements GC {
     @Uninterruptible(reason = "Required by called JavaStackWalker methods. We are at a safepoint during GC, so it does not change anything for this method.", calleeMustBe = false)
     private void blackenStackRoots() {
         Timer blackenStackRootsTimer = timers.blackenStackRoots.open();
+        RuntimeCodeInfoAccess.acquireThreadWriteAccess();
         try {
-            Pointer sp = readCallerStackPointer();
-            CodePointer ip = readReturnAddress();
+            try {
+                Pointer sp = readCallerStackPointer();
+                CodePointer ip = readReturnAddress();
 
-            JavaStackWalk walk = StackValue.get(JavaStackWalk.class);
-            JavaStackWalker.initWalk(walk, sp, ip);
-            walkStack(walk);
+                JavaStackWalk walk = StackValue.get(JavaStackWalk.class);
+                JavaStackWalker.initWalk(walk, sp, ip);
+                walkStack(walk);
 
-            if (SubstrateOptions.MultiThreaded.getValue()) {
-                /*
-                 * Scan the stacks of all the threads. Other threads will be blocked at a safepoint
-                 * (or in native code) so they will each have a JavaFrameAnchor in their VMThread.
-                 */
-                for (IsolateThread vmThread = VMThreads.firstThread(); vmThread.isNonNull(); vmThread = VMThreads.nextThread(vmThread)) {
-                    if (vmThread == CurrentIsolate.getCurrentThread()) {
-                        /*
-                         * The current thread is already scanned by code above, so we do not have to
-                         * do anything for it here. It might have a JavaFrameAnchor from earlier
-                         * Java-to-C transitions, but certainly not at the top of the stack since it
-                         * is running this code, so just this scan would be incomplete.
-                         */
-                        continue;
-                    }
-                    if (JavaStackWalker.initWalk(walk, vmThread)) {
-                        walkStack(walk);
+                if (SubstrateOptions.MultiThreaded.getValue()) {
+                    /*
+                     * Scan the stacks of all the threads. Other threads will be blocked at a
+                     * safepoint (or in native code) so they will each have a JavaFrameAnchor in
+                     * their VMThread.
+                     */
+                    for (IsolateThread vmThread = VMThreads.firstThread(); vmThread.isNonNull(); vmThread = VMThreads.nextThread(vmThread)) {
+                        if (vmThread == CurrentIsolate.getCurrentThread()) {
+                            /*
+                             * The current thread is already scanned by code above, so we do not
+                             * have to do anything for it here. It might have a JavaFrameAnchor from
+                             * earlier Java-to-C transitions, but certainly not at the top of the
+                             * stack since it is running this code, so just this scan would be
+                             * incomplete.
+                             */
+                            continue;
+                        }
+                        if (JavaStackWalker.initWalk(walk, vmThread)) {
+                            walkStack(walk);
+                        }
                     }
                 }
+            } finally {
+                RuntimeCodeInfoAccess.releaseThreadWriteAccess();
             }
         } finally {
             blackenStackRootsTimer.close();
