+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/code/RuntimeCodeInfoMemory.java
@@ -203,18 +203,23 @@ public class RuntimeCodeInfoMemory {
         assert VMOperation.isGCInProgress() : "otherwise, we would need to make sure that the CodeInfo is not freeded by the GC";
         if (table.isNonNull()) {
             int length = NonmovableArrays.lengthOf(table);
-            for (int i = 0; i < length;) {
-                UntetheredCodeInfo info = NonmovableArrays.getWord(table, i);
-                if (info.isNonNull()) {
-                    visitor.visitCode(CodeInfoAccess.convert(info));
-                }
+            RuntimeCodeInfoAccess.acquireThreadWriteAccess();
+            try {
+                for (int i = 0; i < length;) {
+                    UntetheredCodeInfo info = NonmovableArrays.getWord(table, i);
+                    if (info.isNonNull()) {
+                        visitor.visitCode(CodeInfoAccess.convert(info));
+                    }
 
-                // If the visitor removed the current entry from the table, then it is necessary to
-                // visit the now updated entry one more time. However, this could have the effect
-                // that some entries are visited more than once.
-                if (info == NonmovableArrays.getWord(table, i)) {
-                    i++;
+                    // If the visitor removed the current entry from the table, then it is necessary
+                    // to visit the now updated entry one more time. However, this could have the
+                    // effect that some entries are visited more than once.
+                    if (info == NonmovableArrays.getWord(table, i)) {
+                        i++;
+                    }
                 }
+            } finally {
+                RuntimeCodeInfoAccess.releaseThreadWriteAccess();
             }
         }
         return true;
