+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/os/VirtualMemoryProvider.java
@@ -24,6 +24,7 @@
  */
 package com.oracle.svm.core.os;
 
+import com.oracle.svm.core.annotate.Uninterruptible;
 import org.graalvm.compiler.api.replacements.Fold;
 import org.graalvm.nativeimage.ImageSingletons;
 import org.graalvm.word.Pointer;
@@ -83,10 +84,19 @@ public interface VirtualMemoryProvider {
      *            to a multiple of the {@linkplain #getGranularity() granularity}. This value must
      *            not be 0.
      * @param alignment The alignment in bytes of the start of the address range to be reserved.
+     * @param executable Indicates if memory for code is requested.
      * @return An {@linkplain #getAlignment aligned} pointer to the beginning of the reserved
      *         address range, or {@link WordFactory#nullPointer()} in case of an error.
      */
-    Pointer reserve(UnsignedWord nbytes, UnsignedWord alignment);
+    default Pointer reserve(UnsignedWord nbytes, UnsignedWord alignment, boolean executable) {
+        // GR-36766: required for legacy code, remove as soon as no longer needed
+        return reserve(nbytes, alignment);
+    }
+
+    @Uninterruptible(reason = "May be called from uninterruptible code.", mayBeInlined = true)
+    default Pointer reserve(UnsignedWord nbytes, UnsignedWord alignment) {
+        return reserve(nbytes, alignment, false);
+    }
 
     /**
      * Map a region of an open file to the specified address range. When {@linkplain Access#WRITE
@@ -183,4 +193,15 @@ public interface VirtualMemoryProvider {
      * @return 0 when successful, or a non-zero implementation-specific error code.
      */
     int free(PointerBase start, UnsignedWord nbytes);
+
+    /**
+     * Toggle a thread-local flag that tells the OS our intention about the usage of pages that are
+     * mapped with MAP_JIT. Only applicable on macOS. Only enforced on AArch64.
+     *
+     * @param protect If write protection is enabled, pages mapped with MAP_JIT have effectively
+     *            READ|EXEC permissions for the calling thread. If disabled, the permissions for the
+     *            same pages turn to READ|WRITE.
+     */
+    default void jitWriteProtect(boolean protect) {
+    }
 }
